classifier/pyproject.toml
[project]
name = "hanzi-classifier"
version = "0.1.0"
description = "Classifier codebase"
requires-python = ">=3.12,<3.13"

dependencies = [
  "fastmcp",
  "pydantic<2.12",
  "torch",
  "torchvision",
  "torchaudio",
  "numpy",
  "pandas",
  "pillow",
  "opencv-python",
  "scikit-learn",
  "fastapi",
  "jupyter",
  "streamlit",
  "matplotlib",
  "paddleocr>=3.4.0",
  "paddlepaddle>=3.3.0",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }
torchaudio = { index = "pytorch-cpu" }
